Use k-fold cross-validation rather than random samples
k-fold cross-validation ensure that all data is used as both
training and test data, which I think should give better results
than random sampling.

This commit also includes an "extract" method on RDDs that is used
to implement the split into train and test data for k-fold
validation, and tests for this method.