Move static methods to a companion object
Some of the methods on Node don't make use are implemented in
terms of values obtainable from other methods. These have versions
which take this information as a parameter, i.e., which don't
depend on the specific Node object. Thus they are conceptually
static, and have been moved to the companion object.