Factor out parts of tree distance calculation
Calculating distance between two forests was previously a local
function within the tree distance calculation, but is needed
for both distance and mapping calculations, so I've factored it
out into a separate function.

Converting a tree to a list of nodes and descendants was
previously part of the wrapper for treeDistanceMatrix, but is now
a separate function.