Make Node a trait
It makes more sense to make Node a trait, so that users can
provide wrappers for their application-specific node
representations. This makes very little difference to the code;
the trivial Node implementation which just stores a label and
children has been factored out to a SimpleNode class in the test
suite.