Store pending requests
Store is supposed to be a concurrent store, so should deal in a
sensible way with the case when multiple clients are attempting to
process requests. It should attempt to avoid the sitution where
multiple clients both process the same request, although it should
also deal with that situation gracefully. To do that, it allows
clients to claim a particular request, which marks that request as
pending, and removes it from the list of outstanding requests. In
order to deal with the possibility that a process claims a request
but then fails to process it, pending requests are moved back to
the request queue after a (configurable) time limit if they are
not processed.