Split components into separate modules
The trainer uses Spark, and so depends on the Spark libraries
(and their dependencies) but excludes these libraries from its
built artifact (because they will be provided by the Spark server).
On the other hand, the analysis component does not use Spark, but
needs to specifically include the Breeze maths library that is
used by Intel's IMLLIB. But the trainer should not include that
library, because it will be provided by the Spark server.

So, the trainer and the analyser should be compiled separately.
This commit accomplished that split, as well as moving the shared
functionality (the model) into a third module, that both training
and analysis can depend on.

This commit also updates the reference to the IMLLIB submodule to
incorporate a change that allows for that module to be used
without compiling against the Spark libraries, when it's Spark-
related functionality is not used (i.e., in the analysis module).