Analysis service configuration
Rather than hardcoding URLs for the model and the coordinator, it
is better to supply them via a configuration file and/or the
environment. This commit adds a configuration file that refers
to environment variables.

This commit also changes the analysis service to read the model
from a single file, rather than splitting it into two files, and
alters the trainer to write the model in this format.