Add routes for coordinator
Includes routes for the API requests, and routes for the static
parts, i.e. the HTML and JavaScript built by the interface
component. Also includes a symbolic link to the output of the
interface component, so that these files are accessible to the
Play server (and get included in generated artifacts like Docker
containers), and includes an sbt script to run the npm build
process to produce these static files.