Fix error reporting in analysis service
An error in parsing individual documents is supposed to be reported
differently than too few (or too many) files being uploaded, but
these cases were not properly differentiated.

In order to allow better reporting of this error, this commit
changes Protocol.multipartToFiles to return the filename of each
file in the multipart data.

This commit also includes an additional test to verify this
behaviour.